Split package objects into files
