+++ core/src/main/scala/io/finch/response/package.scala
@@ -26,82 +26,9 @@ package io.finch
 
 import com.twitter.finagle.Service
 import com.twitter.finagle.httpx.path.Path
-import com.twitter.finagle.httpx.{Cookie, Response, Status}
+import com.twitter.finagle.httpx.Status
 
 package object response {
-
-  /**
-   * A companion factory object for ''HttpResponse''.
-   *
-   * @param status the http response status
-   * @param headers the HTTP headers map
-   */
-  case class ResponseBuilder(
-      status: Status,
-      headers: Map[String, String] = Map.empty,
-      cookies: Seq[Cookie] = Seq.empty) {
-
-    /**
-     * Creates a new respond with given ''headers''.
-     *
-     * @param headers the HTTP headers map
-     *
-     * @return a respond with headers
-     */
-    def withHeaders(headers: (String, String)*) = copy(headers = this.headers ++ headers)
-
-    /**
-     * Create a new ResponseBuilder with the given ''cookies''.
-     * @param cookies The ''Cookie''s to add to the response
-     * @return a ResponseBuilder with the cookies
-     */
-    def withCookies(cookies: Cookie*) = copy(cookies = this.cookies ++ cookies)
-
-    /**
-     * Creates a ''text/plain'' http response.
-     *
-     * @param plain the response body
-     *
-     * @return a plain text http response
-     */
-    def apply(plain: String) = {
-      val rep = Response(status)
-      rep.setContentString(plain)
-      headers.foreach { case (k, v) => rep.headerMap.add(k, v) }
-      cookies.foreach { rep.addCookie }
-
-      rep
-    }
-
-    /**
-     * Creates an http response with content-type according to the implicit encode.
-     *
-     * @param body the response body
-     * @return an http response
-     */
-    def apply[A](body: A)(implicit encode: EncodeResponse[A]) = {
-      val rep = Response(status)
-      rep.setContentType(encode.contentType)
-      rep.setContentString(encode(body))
-      headers.foreach { case (k, v) => rep.headerMap.add(k, v) }
-      cookies.foreach { rep.addCookie }
-
-      rep
-    }
-
-    /**
-     * Creates an empty http response.
-     *
-     * @return an empty http response
-     */
-    def apply() = {
-      val rep = Response(status)
-      headers.foreach { case (k, v) => rep.headerMap.add(k, v) }
-      cookies.foreach { rep.addCookie }
-
-      rep
-    }
-  }
 
   // 1xx
   object Continue extends ResponseBuilder(Status.Continue)                       // 100
